+++ src/ua/data_types/node_attributes.rs
@@ -1,6 +1,6 @@
 mod variable_attributes;
 
-use open62541_sys::UA_NodeAttributes;
+use open62541_sys::{UA_DataType, UA_NodeAttributes};
 
 use crate::{ua, DataType as _};
 
@@ -9,9 +9,27 @@ crate::data_type!(NodeAttributes);
 macro_rules! derived {
     ($( $name:ident ),* $(,)?) => {
         $(
-            $crate::data_type!($name);
+            paste::paste! {
+                $crate::data_type!([<$name Attributes>]);
+            }
 
-            impl $crate::AsNodeAttributes for $name {
+            impl $crate::Attributes for paste::paste!{[<$name Attributes>]} {
+                fn node_class(&self) -> ua::NodeClass {
+                    paste::paste! {
+                        ua::NodeClass::[<$name:upper>]
+                    }
+                }
+
+                fn attribute_type(&self) -> *const UA_DataType {
+                    <Self as $crate::DataType>::data_type()
+                }
+
+                fn with_display_name(mut self, display_name: &ua::LocalizedText) -> Self {
+                    display_name.clone_into_raw(&mut self.0.displayName);
+                    self.0.specifiedAttributes |= ua::SpecifiedAttributes::DISPLAYNAME.as_u32();
+                    self
+                }
+
                 #[allow(dead_code)]
                 fn as_node_attributes(&self) -> &ua::NodeAttributes {
                     // SAFETY: This transmutes from `Self` to `UA_NodeAttributes`, a strict subset of
@@ -23,10 +41,10 @@ macro_rules! derived {
                 }
             }
 
-            impl Default for $name {
+            impl Default for paste::paste!{[<$name Attributes>]} {
                 fn default() -> Self {
                     paste::paste! {
-                        Self::clone_raw(unsafe { &open62541_sys::[<UA_ $name _default>] })
+                        Self::clone_raw(unsafe { &open62541_sys::[<UA_ $name Attributes_default>] })
                     }
                 }
             }
@@ -37,13 +55,13 @@ macro_rules! derived {
 // This adds basic declarations and shared functionality such as upcasting to `ua::NodeAttributes`.
 // See sub-modules for type-specific implementations, e.g. `variable_attributes`.
 derived!(
-    ObjectAttributes,
-    VariableAttributes,
-    MethodAttributes,
-    ObjectTypeAttributes,
-    VariableTypeAttributes,
-    ReferenceTypeAttributes,
-    DataTypeAttributes,
-    ViewAttributes,
-    // GenericAttributes, // Omitted for now because the `Default` impl above cannot be used here.
+    Object,
+    Variable,
+    Method,
+    ObjectType,
+    VariableType,
+    ReferenceType,
+    DataType,
+    View,
+    // Generic, // Omitted for now because the `Default` impl above cannot be used here.
 );